# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teleasistencia Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categorias": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Obtener todas las categorías",
                "operationId": "7dab55b7d8cf8c6edabd96c94524f5a1",
                "responses": {
                    "200": {
                        "description": "Lista de categorías recuperada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CategoriaResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron categorías"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/categorias/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Obtener detalles de una categoría",
                "operationId": "23b331d44668d58cda020df86ac2cf8c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría a recuperar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la categoría recuperados con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CategoriaResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CategoriaResource": {
                "description": "Esquema del recurso Categoría",
                "properties": {
                    "id": {
                        "description": "Identificador de la categoría",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre de la categoría",
                        "type": "string",
                        "example": "Urgencias Médicas"
                    },
                    "descripcion": {
                        "description": "Descripción de la categoría",
                        "type": "string",
                        "example": "Categoría para gestionar urgencias médicas prioritarias",
                        "nullable": true
                    },
                    "subcategorias": {
                        "description": "Lista de subcategorías asociadas a la categoría",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SubcategoriaResource"
                        }
                    },
                    "created_at": {
                        "description": "Fecha de creación de la categoría",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-23 12:00:00"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización de la categoría",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24 12:00:00"
                    }
                },
                "type": "object"
            },
...
            "SubcategoriaResource": {
                "description": "Esquema del recurso Subcategoría",
                "properties": {
                    "id": {
                        "description": "Identificador único de la subcategoría",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre de la subcategoría",
                        "type": "string",
                        "example": "Emergencia leve"
                    },
                    "categoria_id": {
                        "$ref": "#/components/schemas/CategoriaResource"
                    },
                    "created_at": {
                        "description": "Fecha de creación de la subcategoría",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-23 12:00:00"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización de la subcategoría",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-24 12:00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}